+++ src/config.h
@@ -118,6 +118,7 @@ class ConfigMain : public ConfigFile {
     const char * wifi_ssid;
     const char * wifi_password;
     uint16_t     wifi_restart_core_sec;
+    uint16_t     wifi_publish_min_interval_sec;
 
     /* Station settings */
     const char * station_name;